fix(dp): fix circular import - #300
