+++ src/position-slicer.ts
@@ -3,16 +3,17 @@
 // and display their slices
 
 import * as vscode from 'vscode';
-import { getConfig, getFlowrSession, isVerbose, updateStatusBar } from './extension';
+import { getFlowrSession, updateStatusBar } from './extension';
 import { makeUri, getReconstructionContentProvider, showUri } from './doc-provider';
 import { getPositionAt, makeSlicingCriteria } from './flowr/utils';
 import type { DecoTypes } from './slice';
 import { displaySlice, makeSliceDecorationTypes } from './slice';
 import { getSelectionSlicer } from './selection-slicer';
-import { Settings } from './settings';
+import { Settings , getConfig, isVerbose } from './settings';
 import { SliceDirection } from '@eagleoutice/flowr/core/steps/all/static-slicing/00-slice';
 import type { NodeId } from '@eagleoutice/flowr/r-bridge/lang-4.x/ast/model/processing/node-id';
 import type { SourceRange } from '@eagleoutice/flowr/util/range';
+
 
 const positionSlicerAuthority = 'doc-slicer';
 const positionSlicerSuffix = 'Slice';
