+++ src/slice.ts
@@ -1,8 +1,7 @@
 import * as vscode from 'vscode';
 import type { SourceRange } from '@eagleoutice/flowr/util/range';
-import { getConfig, isVerbose } from './extension';
 import type { SliceDisplay } from './settings';
-import { Settings } from './settings';
+import { getConfig, isVerbose, Settings } from './settings';
 import { getSelectionSlicer, showSelectionSliceInEditor } from './selection-slicer';
 import { disposeActivePositionSlicer, getActivePositionSlicer, addCurrentPositions, positionSlicers } from './position-slicer';
 import type { NodeId } from '@eagleoutice/flowr/r-bridge/lang-4.x/ast/model/processing/node-id';
